Do not include main project build in testing projects
The tests do use different Java and Gradle versions to make sure the
plugin can be applied using those. However the main build uses APIs that
are not available on all Gradle versions, leading to build errors.
Anyway, including the main project is not required, so it can be easily
removed to avoid that problem.